testing an animation part change
OnHide is doing SetSequence(0) which makes sense in a single animation
on/off context but I think doesn't make sense in a context with multiple
events controlling different animations. I'm adding a "Reset On Hide"
option (default true) that keeps this functionality, but when it is set
to false, the model isn't automatically t-posed when an animation part
is hidden.