Add pace tool "square model scales"
Adds a pace tool that will square the scales of all parts who have a
Model field selectively if that field contains a given string. This tool
will help repair costumes that were created between Apr 04 2014 and July
08 2014 that did not use legacy scale. A filter is being used to select
which models to scale because not all models will need to be scaled back
(in many cases, only default.mdl will be- which is why it is the default
string).