Add mdl_import part
Adds "mdl_import" part that lets users download mdl files from remote
hosts. To use it, create a model part and make a mdl_import part as a
child of that model part. Then, paste a URL in the mdl_import part.

Be aware that the mdl,vvd,phy, and vtx files should all be available at
the same URL with different suffixes (eg. example.com/house/house.mdl,
example.com/house/house.dx80.vtx, example.com/house/house.vvd... and so
on). For this reason, filehosting services that obfuscate the filename
(eg dropbox, mediafire) are NOT feasible for use with mdl_import.