more vfs icon fixing stuff
client should request the vfs status of other clients on join. clients
shouldn't set their own status until they've joined.

if anyone knows of a clientside hook equivalent to PlayerInitialSpawn
(args not needed) so that I don't need to reflect the serverside hook to
the player like this through net it would be much appreciated to know