Fix flickering / improve draw suppression
The most hax, ever.

Fixes the frame counting / suppression of PostDrawOpaqueRenderable calls. The frame suppression no longer depends on how many draw calls are in the current frame, but matters that the render stack executed in a correct order (There's no reason why it shouldn't). 

This fixes outfit flickering. 

There's a pull request in on the master repo of GMod to hopefully make this less hacky .

https://github.com/Facepunch/garrysmod/pull/1406