feat(select): Allow passing in outside events
This is done so that it is possible to redirect input from DModelPanels
to allow using the Select feature inside of them, similar to how it is
done in mctrl.lua. The events were kept as hooks instead of embedding
them as part of the editor (as done in mctrl) since it's not an integral
feature (as mctrl is).