A DFrame window to contain the zoom slider (Alternative to Yagira's slider)
It's not good that the zoom slider is just invisible in the corner, and even the rare few who know it exists have difficulty because of how broken its positioning is (the slider pin is half out of the screen, and it is also not easy to target because it is invisible and small). A window has some advantages over a slider that is always at the bottom of the screen.