popup fixes, add commands in menu bar
fix pace being localized as nil in base_part. we need it in the popups setup

in the popups code it should be "menu bar" and not "editor bar"

add command events, sequenced commands and command proxies to the player tab of the menu bar. the menus will force-refresh to auto-update according to current values