changes to Load VMT outputs
moved the note-taking to the part code. it only applied to the model's existing VMT quickloads. now it should work with the asset browser and material pastes

new verbosity convar for the console prints, add descriptive titles and space out the prints a bit more.

when erroring, apply a warning circle for each VMT line that errored