better way to handle bulk select in nested setups
pac_bulk_select_subsume is a new option to toggle the behavior where selecting a part deselects its children and prevents selecting those children thereafter

having it off will allow operations on arbitrary parts in nested setups, but it can still cause issues with "broad" operations if you, for example, copy a series of parts that are parents, grandparents, great-grandparents etc. subsuming would've prevented excess copies of children. or if you move / insert, the hierarchy can break. it'll make sense if you think about how it works. bulk select does single operations one by one on a list

use flashnotification when doing bulk selections and deselections, in keeping with the general principle of explaining what's going on when editor actions are being done

with that, also automatically add the bulk select partmenu option at the end, when there's a bulk selection and it's not already part of the menu. it's not a perfect solution to prevent confusion but it might be helpful for people when they accidentally discover bulk select.