more camera part quick setups (place camera from editor view)
apply editor view - translates editor camera position into local space, with a series of sub-options to switch between editor or default camera FOV, and select bones

centralized the zoom slider's bounds-setting functions into one
Add an option on the fov slider to raise its fov limit. By default it should stay at 100, but expanding it to 179 would be used to mirror the camera part's limits to set up some of the more exaggerated camera shots we might see in stylized animation and such (e.g. high FOV close-up on a weapon to make it look bigger).

converted some of the camera quick setup options into right clickable options that don't automatically close the menu, allowing a better experience in trying out different configs. I'm experimenting with that, would be good when there are occasionally options with multiple choices which would benefit from immediate 1-click feedback. I will tentatively explore this, but it should be used sparingly, for various reasons. misusing it might result in anti-QoL because most options introduce side effects that don't get undone by neighboring options, since they often do separate things

try to look for an alternate head if the standard valvebiped head isn't found

the base height used for calculation is now a traced hit to sample the top of the head instead of the base of the head bone. hopefully it works better. it's also manually adjustable (right clickable options refresh the height variable) to other alternates (between head bone's base, trace, hull height, eyepos), it's ultimately a micro difference but will be added for the sake of information and comparison

add option to clone the camera position as a dummy model for interpolators